# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "x402-github-bounty",
    "version": "0.1.0",
    "description": "Non-custodial GitHub bounties. Minting a certificate live-verifies the issue through the GitHub REST API and returns an HMAC-signed certificate in the 200 body; a verification report is built from live GitHub state at request time. Paid routes speak x402: an unpaid request returns 402 with a dual-rail `accepts` array (USDC on Base and USDC on Solana) and the client picks the rail. The service never holds bounty funds.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "summary": "Fund GitHub issues with x402 — signed bounty certificates, merged-PR verification reports, settlement receipts.",
    "contact": {
      "name": "nirholas",
      "email": "nichxbt@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4021",
      "description": "local self-host"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Service info (free)",
        "responses": {
          "200": {
            "description": "Metadata"
          }
        },
        "operationId": "getServiceCard"
      }
    },
    "/health": {
      "get": {
        "summary": "Health (free)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "operationId": "getHealth"
      }
    },
    "/bounties": {
      "get": {
        "summary": "Public bounty board (free)",
        "responses": {
          "200": {
            "description": "All bounty records (secrets omitted)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer"
                    },
                    "bounties": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BountyPublic"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "operationId": "listBounties"
      },
      "post": {
        "summary": "Create a signed bounty certificate ($0.01 via x402)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "issueUrl",
                  "amount"
                ],
                "properties": {
                  "issueUrl": {
                    "type": "string",
                    "example": "https://github.com/owner/repo/issues/123"
                  },
                  "amount": {
                    "type": "number",
                    "exclusiveMinimum": 0
                  },
                  "terms": {
                    "type": "string"
                  },
                  "funder": {
                    "type": "string"
                  },
                  "expiryDays": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 365
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed bounty certificate + settleKey (the purchased artifact)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BountyCreated"
                }
              }
            },
            "headers": {
              "X-PAYMENT-RESPONSE": {
                "description": "Base64 JSON settlement receipt (rail, network, transaction, payer, amount).",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "404": {
            "description": "Issue not found on GitHub"
          },
          "409": {
            "description": "Issue already closed"
          },
          "502": {
            "description": "GitHub upstream error"
          }
        },
        "operationId": "createBounty",
        "description": "Price $0.01 in USDC. Pay on Base or Solana — the 402 lists both rails. The signed certificate is the purchased artifact and is returned in this response."
      }
    },
    "/verify/{bountyId}": {
      "get": {
        "summary": "Live merged-PR verification report ($0.002 via x402)",
        "parameters": [
          {
            "name": "bountyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signed verification report from live GitHub state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationReport"
                }
              }
            },
            "headers": {
              "X-PAYMENT-RESPONSE": {
                "description": "Base64 JSON settlement receipt (rail, network, transaction, payer, amount).",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "404": {
            "description": "Unknown bountyId"
          },
          "502": {
            "description": "GitHub upstream error"
          }
        },
        "operationId": "verifyBounty",
        "description": "Price $0.002 in USDC. Pay on Base or Solana — the 402 lists both rails."
      }
    },
    "/settle/{bountyId}": {
      "post": {
        "summary": "Settle a bounty (free, settle-key auth)",
        "parameters": [
          {
            "name": "bountyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "settleKey",
                  "payoutAddress"
                ],
                "properties": {
                  "settleKey": {
                    "type": "string"
                  },
                  "payoutAddress": {
                    "type": "string",
                    "pattern": "^0x[0-9a-fA-F]{40}$"
                  },
                  "prNumber": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed payout receipt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutReceipt"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "description": "settleKey mismatch"
          },
          "404": {
            "description": "Unknown bountyId"
          },
          "409": {
            "description": "Bounty already settled/expired"
          }
        },
        "operationId": "settleBounty",
        "description": "Free. Authorised by the settleKey issued alongside the certificate; returns the signed payout receipt immediately."
      }
    },
    "/.well-known/x402": {
      "get": {
        "operationId": "getX402Manifest",
        "summary": "x402 resource manifest (free)",
        "responses": {
          "200": {
            "description": "Machine-readable listing of paid resources, prices and rails.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/skill.md": {
      "get": {
        "operationId": "getSkillFile",
        "summary": "Agent-facing skill file (free)",
        "responses": {
          "200": {
            "description": "skill.md",
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/check-signature": {
      "post": {
        "operationId": "checkSignature",
        "summary": "Validate a signed artifact (free)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "payload",
                  "signature"
                ],
                "properties": {
                  "payload": {
                    "type": "object",
                    "description": "The signed object exactly as returned."
                  },
                  "signature": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "type": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "checkedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "note": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BountyPublic": {
        "type": "object",
        "properties": {
          "bountyId": {
            "type": "string"
          },
          "issueUrl": {
            "type": "string"
          },
          "issueTitle": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "settled",
              "expired"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          }
        }
      },
      "BountyCreated": {
        "type": "object",
        "properties": {
          "certificate": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "bountyId": {
                "type": "string"
              },
              "issueUrl": {
                "type": "string"
              },
              "repo": {
                "type": "string"
              },
              "issueNumber": {
                "type": "integer"
              },
              "issueTitle": {
                "type": "string"
              },
              "amount": {
                "type": "string"
              },
              "currency": {
                "type": "string"
              },
              "terms": {
                "type": "string"
              },
              "funder": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "expiresAt": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          "signature": {
            "type": "string"
          },
          "settleKey": {
            "type": "string"
          },
          "settleKeyNote": {
            "type": "string"
          },
          "verifyUrl": {
            "type": "string"
          },
          "algorithm": {
            "type": "string"
          },
          "receipt": {
            "type": [
              "object",
              "null"
            ]
          }
        }
      },
      "VerificationReport": {
        "type": "object",
        "properties": {
          "report": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "bountyId": {
                "type": "string"
              },
              "issueUrl": {
                "type": "string"
              },
              "bountyStatus": {
                "type": "string"
              },
              "amount": {
                "type": "string"
              },
              "issue": {
                "type": "object",
                "properties": {
                  "state": {
                    "type": "string"
                  },
                  "closedAt": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "stateReason": {
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                }
              },
              "mergedPrs": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "number": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "author": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "url": {
                      "type": "string"
                    },
                    "mergedAt": {
                      "type": "string"
                    },
                    "mergeCommit": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  }
                }
              },
              "eligible": {
                "type": "boolean"
              },
              "eligibleReason": {
                "type": "string"
              },
              "source": {
                "type": "string"
              },
              "checkedAt": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          "signature": {
            "type": "string"
          },
          "algorithm": {
            "type": "string"
          },
          "receipt": {
            "type": [
              "object",
              "null"
            ]
          }
        }
      },
      "PayoutReceipt": {
        "type": "object",
        "properties": {
          "receipt": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "bountyId": {
                "type": "string"
              },
              "issueUrl": {
                "type": "string"
              },
              "amount": {
                "type": "string"
              },
              "currency": {
                "type": "string"
              },
              "payoutAddress": {
                "type": "string"
              },
              "payoutPr": {
                "type": [
                  "integer",
                  "null"
                ]
              },
              "settledAt": {
                "type": "string",
                "format": "date-time"
              },
              "note": {
                "type": "string"
              }
            }
          },
          "signature": {
            "type": "string"
          }
        }
      },
      "PaymentRequirements": {
        "type": "object",
        "description": "One accepted way to pay for this resource (one rail).",
        "required": [
          "scheme",
          "network",
          "maxAmountRequired",
          "resource",
          "description",
          "mimeType",
          "payTo",
          "maxTimeoutSeconds",
          "asset"
        ],
        "properties": {
          "scheme": {
            "type": "string",
            "enum": [
              "exact"
            ]
          },
          "network": {
            "type": "string",
            "enum": [
              "base",
              "base-sepolia",
              "solana",
              "solana-devnet"
            ]
          },
          "maxAmountRequired": {
            "type": "string",
            "description": "Amount in asset base units (USDC has 6 decimals)."
          },
          "resource": {
            "type": "string",
            "format": "uri"
          },
          "description": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "payTo": {
            "type": "string"
          },
          "maxTimeoutSeconds": {
            "type": "integer"
          },
          "asset": {
            "type": "string",
            "description": "USDC contract address or SPL mint."
          },
          "extra": {
            "type": "object"
          }
        }
      },
      "PaymentRequiredBody": {
        "type": "object",
        "required": [
          "x402Version",
          "accepts"
        ],
        "properties": {
          "x402Version": {
            "type": "integer",
            "enum": [
              1
            ]
          },
          "error": {
            "type": "string"
          },
          "accepts": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/PaymentRequirements"
            }
          }
        }
      }
    },
    "responses": {
      "PaymentRequired": {
        "description": "Payment required. The `accepts` array lists both rails — USDC on Base and USDC on Solana. Sign one, base64 it into `X-PAYMENT`, retry.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PaymentRequiredBody"
            },
            "example": {
              "x402Version": 1,
              "error": "X-PAYMENT header is required",
              "accepts": [
                {
                  "scheme": "exact",
                  "network": "base-sepolia",
                  "maxAmountRequired": "10000",
                  "resource": "http://localhost:4021/bounties",
                  "description": "Mint a signed bounty certificate against a live-verified open GitHub issue",
                  "mimeType": "application/json",
                  "payTo": "0x40252CFDF8B20Ed757D61ff157719F33Ec332402",
                  "maxTimeoutSeconds": 60,
                  "asset": "0x036CbD53842c5426634e7929541eC2318f3dCF7e",
                  "extra": {
                    "name": "USDC",
                    "version": "2"
                  }
                },
                {
                  "scheme": "exact",
                  "network": "solana",
                  "maxAmountRequired": "10000",
                  "resource": "http://localhost:4021/bounties",
                  "description": "Mint a signed bounty certificate against a live-verified open GitHub issue",
                  "mimeType": "application/json",
                  "payTo": "WwwuGbqHrwF5RG89KhUbmRWEvjnRH9k5kVM5p7T3WwW",
                  "maxTimeoutSeconds": 60,
                  "asset": "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v",
                  "extra": {
                    "name": "USD Coin",
                    "decimals": 6,
                    "feePayer": "2wKupLR9q6wXYppw8Gr2NvWxKBUqm4PPJKkQfoxHDBg4"
                  }
                }
              ]
            }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid input",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "externalDocs": {
    "description": "Docs and tutorial",
    "url": "https://nirholas.github.io/x402-github-bounty/"
  }
}